+++ src/app/ng-unit-test-helpers.ts
@@ -1,9 +1,8 @@
 // Helper functions for angular unit tests.
 
-import { LhcDataService } from '../lib/lhc-data.service';
 import { TestBed } from '@angular/core/testing';
+import {InternalUtil} from "./../lib/lforms/internal-utils.js";
 
-let lhcDataService;
 
 /**
  *  Returns the element with the ID constructed for the given item and answer.
@@ -12,6 +11,7 @@ let lhcDataService;
  * @param item an item from an LForms items array
  * @param answer an answer from the item's answers array
  */
+export function getItemAnswerElem(element, item, answer) {
  return element.querySelector('#' +
-    lhcDataService.getItemAnswerId(item, answer).replaceAll('|', '\\|'));
+    InternalUtil.getItemAnswerId(item, answer).replaceAll('|', '\\|'));
 }