+++ src/lib/lforms/lhc-form-data.ts
@@ -9,6 +9,7 @@ import LhcFormUtils from "./lhc-form-utils.js";
 import CommonUtils from "./lhc-common-utils.js";
 import {InternalUtil} from "./internal-utils.js";
 import version from '../../version.json';
+import language from '../../../language-config.json';
 
 import Validation from "./lhc-form-validation.js"
 
@@ -58,6 +59,9 @@ export default class LhcFormData {
   // active item, where a input field in the row has the focus
   _activeItem = null;
 
+  // How many levels of items the form has.
+  _totalLevelsOfHierarchy = 0;
+
   // default template options
   _defaultTemplateOptions = {
     // whether question code is displayed next to the question
@@ -91,11 +95,11 @@ export default class LhcFormData {
       }
     },
     // whether to hide tree line styles
-    hideTreeLine: false,
+    hideTreeLine: 'auto',
     // whether to hide indentation before each item
     hideIndentation: false,
     // whether to hide repetition numbers next to the item's text
-    hideRepetitionNumber: false,
+    hideRepetitionNumber: true,
     // whether to display score along with text when there scores in answers
     displayScoreWithAnswerText: true,
     // whether to show the filtered html content from the rendering-xhtml extension
@@ -938,6 +942,7 @@ export default class LhcFormData {
     this._codePath = "";
     this._idPath = "";
     this._displayLevel = 0;
+    this._totalLevelsOfHierarchy = 0;
     this._activeItem = null;
 
     // template
@@ -1277,6 +1282,9 @@ export default class LhcFormData {
       item._elementId = item.linkId.replaceAll('\\', '\\\\')
         .replaceAll(this.PATH_DELIMITER, '\\'+ this.PATH_DELIMITER) + item._idPath;
       item._displayLevel = parentItem._displayLevel + 1;
+      if (item._displayLevel > this._totalLevelsOfHierarchy) {
+        this._totalLevelsOfHierarchy = item._displayLevel;
+      }
 
       // set last sibling status
       item._lastSibling = i === lastSiblingIndex;
@@ -1353,29 +1361,29 @@ export default class LhcFormData {
       else if (item._hasAnswerList) {
         if (item.answerConstraint === "optionsOrString") {
           if (item.externallyDefined)
-            item._placeholder = item._multipleAnswers ? "Search for or type values" : "Search for or type a value";
+            item._placeholder = item._multipleAnswers ? language.placeholderSearchMultiOptionsOrString : language.placeholderSearchSingleOptionsOrString;
           else
-            item._placeholder = item._multipleAnswers ? "Select one or more or type a value" : "Select one or type a value";
+            item._placeholder = item._multipleAnswers ? language.placeholderSelectMultiOptionsOrString : language.placeholderSelectSingleOptionsOrString;
         }
         // INT, ST, DT, TM and CODING (and answerConstraint == 'optionsOnly')
         else {
           if (item.externallyDefined)
-            item._placeholder = item._multipleAnswers ? "Search for values" : "Search for value";
+            item._placeholder = item._multipleAnswers ? language.placeholderSearchMulti : language.placeholderSearchSingle;
           else
-            item._placeholder = item._multipleAnswers ? "Select one or more" : "Select one";
+            item._placeholder = item._multipleAnswers ? language.placeholderSelectMulti : language.placeholderSelectSingle;
         }
       }
       // other types
       else {
         switch (item.dataType) {
           case CONSTANTS.DATA_TYPE.DT:
-            item._placeholder = "MM/DD/YYYY";
+            item._placeholder = language.placeholderDate;
             break;
           case CONSTANTS.DATA_TYPE.DTM:
-            item._placeholder = "MM/DD/YYYY HH:MM:SS";
+            item._placeholder = language.placeholderDateTime;
             break;
           case CONSTANTS.DATA_TYPE.TM:
-            item._placeholder = "HH:MM:SS";
+            item._placeholder = language.placeholderTime;
             break;
           case CONSTANTS.DATA_TYPE.CODING:
             if (!item.answerConstaint || item.answerConstraint === 'optionsOnly' ) {
@@ -1386,9 +1394,9 @@ export default class LhcFormData {
             }
             else if (item.answerConstraint === 'optionsOrString') {
               if (item.externallyDefined)
-              item._placeholder = item._multipleAnswers ? "Search for or type values" : "Search for or type a value";
+              item._placeholder = item._multipleAnswers ? language.placeholderSearchMultiOptionsOrString : language.placeholderSearchSingleOptionsOrString;
             else
-              item._placeholder = item._multipleAnswers ? "Select one or more or type a value" : "Select one or type a value";
+              item._placeholder = item._multipleAnswers ? language.placeholderSelectMultiOptionsOrString : language.placeholderSelectSingleOptionsOrString;
             }
             break;
           case "SECTION":
@@ -1399,10 +1407,10 @@ export default class LhcFormData {
           case CONSTANTS.DATA_TYPE.INT:
           case CONSTANTS.DATA_TYPE.REAL:
           case CONSTANTS.DATA_TYPE.QTY:
-            item._placeholder = "Type a number";
+            item._placeholder = language.placeholderTypeANumber;
             break;
           default: {
-            item._placeholder = "Type a value";
+            item._placeholder = language.placeholderTypeAValue;
           }
         }
       }
@@ -1534,6 +1542,9 @@ export default class LhcFormData {
       item._idPath = parentItem._idPath + this.PATH_DELIMITER + item._id;
       item._elementId = item.linkId + item._idPath;
       item._displayLevel = parentItem._displayLevel + 1;
+      if (item._displayLevel > this._totalLevelsOfHierarchy) {
+        this._totalLevelsOfHierarchy = item._displayLevel;
+      }
       item._parentItem = parentItem;
       item._repeatingSectionList = null;
 
@@ -2266,7 +2277,7 @@ export default class LhcFormData {
 
     this._resetInternalData();
 
-    var readerMsg = 'Added ' + this.itemDescription(item);
+    var readerMsg = language.added + this.itemDescription(item);
     this._actionLogs.push(readerMsg);
 
     // run FHIRPath expression when a new item is added
@@ -2347,7 +2358,7 @@ export default class LhcFormData {
 
     this._resetInternalData();
 
-    var readerMsg = 'Added ' + this.itemDescription(item);
+    var readerMsg = language.added + this.itemDescription(item);
     this._actionLogs.push(readerMsg);
 
     return newItem;
@@ -2519,7 +2530,7 @@ export default class LhcFormData {
     }
 
     this._resetInternalData();
-    var readerMsg = 'Removed ' + this.itemDescription(item);
+    var readerMsg = language.removed + this.itemDescription(item);
     this._actionLogs.push(readerMsg);
 
     // run FHIRPath expression when a new item is removed