+++ src/lib/lhc-data.service.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@angular/core';
 import { ScreenReaderLog } from './screen-reader-log';
 import CommonUtils from "./lforms/lhc-common-utils.js";
 import {InternalUtil} from "./lforms/internal-utils.js";
+import language from '../../language-config.json';
 declare var LForms: any;
 // @Injectable({
 //   providedIn: 'root'
@@ -329,32 +330,45 @@ export class LhcDataService {
 
   /**
    * Get CSS classes for the tree line
+   * @param isTopLevelItem whether it's a top level item
    * @returns {string}
    */
-  getTreeLineClass() {
+  getTreeLineClass(isTopLevelItem = false) {
+    if (isTopLevelItem) {
+      return '';
+    }
     const templateOptions = this.getLhcFormData().templateOptions;
-    return templateOptions.hideTreeLine || templateOptions.hideIndentation ? '' : 'lhc-tree-line';
+    if (templateOptions.hideIndentation ||
+      templateOptions.hideTreeLine === 'true' ||
+      templateOptions.hideTreeLine === true ||
+      (templateOptions.hideTreeLine === 'auto' && this.getLhcFormData()._totalLevelsOfHierarchy <= 3)) {
+      return '';
+    } else {
+      return 'lhc-tree-line';
+    }
   }
 
   /**
    * Get CSS classes for the indentation
+   * @param isTopLevelItem whether it's a top level item
    * @returns {string}
    */
-  getIndentationClass() {
-    return this.getLhcFormData().templateOptions.hideIndentation ? '' : 'lhc-indentation';
+  getIndentationClass(isTopLevelItem = false) {
+    return isTopLevelItem || this.getLhcFormData().templateOptions.hideIndentation ? '' : 'lhc-indentation';
   }
 
   /**
    * get CSS class list for an item
    * @param item an item in a form
    * @param viewMode view mode of the item
+   * @param isTopLevelItem whether item is a top level item
    */
-  getItemClassList(item, viewMode) {
+  getItemClassList(item, viewMode, isTopLevelItem = false) {
     const classList = [
       'lhc-item',
       this.getItemViewModeClass(item, viewMode),
-      this.getTreeLineClass(),
-      this.getIndentationClass(),
+      this.getTreeLineClass(isTopLevelItem),
+      this.getIndentationClass(isTopLevelItem),
       this.getSiblingStatus(item),
       this.getRowClass(item),
       this.getActiveRowClass(item)
@@ -499,8 +513,7 @@ export class LhcDataService {
       anyEmpty = this.lhcFormData.areAnyRepeatingItemsEmpty(item);
       if (anyEmpty && item._showUnusedItemWarning) {
         if (!item._unusedItemWarning)
-          item._unusedItemWarning = 'Please enter info in the blank "' +
-            item._text+'"';
+          item._unusedItemWarning = language.pleaseEnterInfoForTheBlank.replace('{lformsParam}',  item._text);
       }
     }
     if (!anyEmpty) {