+++ test/cypress/support/testUtilFacade.js
@@ -65,12 +65,9 @@ export const TestUtil = {
    * Wait for LForms to be loaded.
    */
   waitForLFormsLoaded() {
-    cy.window().then(win => {
+    cy.window().should(win => {
       let LForms = win.LForms;
-      cy.wrap({1: 2}).should(()=> {
-        // not sure this wrap is needed, but without it sometimes the expect fails.
-        expect(typeof(LForms) === "object").to.be.true;
-      });
+      expect(typeof(LForms) === "object").to.be.true;
     });
   },
 
@@ -79,7 +76,7 @@ export const TestUtil = {
    * Wait for the externally linked FHIR libraries to be loaded.
    */
   waitForFHIRLibsLoaded() {
-    cy.window().then(win => {
+    cy.window().should(win => {
       let LForms = win.LForms;
       expect(typeof(LForms.FHIR) === "object" && typeof(LForms.FHIR.R4) === "object" &&
           typeof(LForms.FHIR.STU3) === "object" && typeof(LForms.FHIR.R5) === "object").to.be.true;